behavioral/command/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
